wifi: rtl8852be: upgrade driver to v1.19.11-175-g5130b5e2d.20240429

+++ drivers/net/wireless/rockchip_wlan/rtl8852be/phl/phl_msg_hub.h
@@ -27,7 +27,7 @@ phl_msg_hub_register_recver(void *phl, struct phl_msg_receiver *ctx,
 			    enum phl_msg_recver_layer layer);
 enum rtw_phl_status
 phl_msg_hub_update_recver_mask(void *phl, enum phl_msg_recver_layer layer,
-			       u8 *mdl_id, u32 len, u8 clr);
+			       u8 *mdl_id, u8 len, u8 clr);
 enum rtw_phl_status
 phl_msg_hub_deregister_recver(void *phl, enum phl_msg_recver_layer layer);
